Update to Version 0.2.0
This updates the crate to version 0.2.0, which replaces how the monitor detection works to rely on Winit instead of our own, hand rolled solution.

The `determine_monitor()` function remains, but has been deprecated as it's unlikely to be useful as you can actually access the underlying Winit Window Data, the deprecation notice informs the user of this option. However, there may be some slight disagreements between the two methods.